纯 Go 跨平台 GUI 框架（参考 Wails v3 架构），Windows 端路线图 GUI-1~4 全部落地：

- 纯 Go WebView2 COM 绑定（无 CGO / WebView2Loader.dll 依赖）
- aluka://app/ 内存虚拟协议（WebResourceRequested 拦截，零 TCP 端口）
- 专用 UI 线程模型 + COM 生命周期（AddRef/WM_CLOSE 路由）
- window.aluka 桥接：RPC / 双向事件 / 原生对话框 / 窗口控制
- 系统托盘 + 原生弹出菜单 + 全局快捷键 + Mica/Acrylic 特效
- 无边框拖拽（--aluka-draggable / -webkit-app-region / data-aluka-drag）
  + 边缘八方向缩放 + 双击最大化
- aluka build --gui 单文件打包：前端资源内嵌 + PE 图标注入 + 免黑框
- Aluka.gui / aluka:gui JS API 全量暴露
- 附带修复：darwin 交叉编译（jit/builtin 平台拆分）、App.Quit 重入死锁、
  app.run 事件循环竞态、PCWSTR GC 悬垂隐患
- demo/studio 全功能演示（dev 运行 + 单文件打包双模式验证）

待办：macOS/Linux 平台层、白屏偶发问题（已定位为遮挡触发的 renderer
无响应，带观测日志与 Reload 自愈，待复现跟进）

+++ internal/runtime/globals/process_platform_unix.go
@@ -22,8 +22,8 @@ func setUmask(mask int) int {
 func getProcessUsage() (user, system int64) {
 	var ru syscall.Rusage
 	if err := syscall.Getrusage(syscall.RUSAGE_SELF, &ru); err == nil {
-		user = ru.Utime.Sec*1e6 + ru.Utime.Usec
-		system = ru.Stime.Sec*1e6 + ru.Stime.Usec
+		user = ru.Utime.Sec*1e6 + int64(ru.Utime.Usec)
+		system = ru.Stime.Sec*1e6 + int64(ru.Stime.Usec)
 	}
 	return
 }
